feat: modularize step two media and character flow
