+++ src/app/[locale]/tell-your-story/step-2/page.tsx
@@ -1,44 +1,18 @@
 'use client';
 
 import { SignedIn, SignedOut, RedirectToSignIn } from '@clerk/nextjs';
-import Image from 'next/image';
 import StepNavigation from '../../../../components/StepNavigation';
-import { useState, useRef, useEffect, useCallback } from 'react';
+import { useState } from 'react';
 import { useRouter } from 'next/navigation';
 import { useTranslations } from 'next-intl';
 import { trackStoryCreation } from '../../../../lib/analytics';
-
-// Writing tips that rotate
-const WRITING_TIPS = [
-  { icon: '🌍', key: 'setting' },
-  { icon: '👥', key: 'characters' },
-  { icon: '⚔️', key: 'conflict' },
-  { icon: '❤️', key: 'emotion' },
-  { icon: '🎭', key: 'twist' },
-  { icon: '🎯', key: 'goal' },
-  { icon: '✨', key: 'magic' }
-];
+import MediaCapture from './MediaCapture';
+import CharacterSelection from './CharacterSelection';
+import WritingTips from './WritingTips';
+import { useStep2Session } from '@/hooks/useStep2Session';
+import { useMediaUpload } from '@/hooks/useMediaUpload';
 
 type ContentType = 'text' | 'images' | 'audio';
-
-interface SessionData {
-  text: string;
-  images: string[];
-  audio: string | null;
-  lastSaved: number;
-}
-
-interface Character {
-  characterId: string;
-  name: string;
-  type?: string;
-  role?: string;
-  age?: string;
-  traits?: string[];
-  characteristics?: string;
-  physicalDescription?: string;
-  photoUrl?: string;
-}
 
 export default function Step2Page() {
   const router = useRouter();
@@ -48,289 +22,25 @@ export default function Step2Page() {
   // Modal states
   const [activeModal, setActiveModal] = useState<ContentType | null>(null);
   
-  // Content states
-  const [storyText, setStoryText] = useState('');
-  const [uploadedImages, setUploadedImages] = useState<Array<{file: File, preview: string}>>([]);
-  const [uploadedAudio, setUploadedAudio] = useState<{file: File, preview: string} | null>(null);
   
   // UI states
-  const [isCapturing, setIsCapturing] = useState(false);
-  const [isRecording, setIsRecording] = useState(false);
   const [isCreatingStory, setIsCreatingStory] = useState(false);
   const [showLoadingModal, setShowLoadingModal] = useState(false);
-  const [currentTipIndex, setCurrentTipIndex] = useState(0);
-  const [isSaving, setIsSaving] = useState(false);
   
   // Character states
-  const [existingCharacters, setExistingCharacters] = useState<Character[]>([]);
   const [selectedCharacterIds, setSelectedCharacterIds] = useState<string[]>([]);
-  const [showCharacterSelection, setShowCharacterSelection] = useState(false);
-  const [isLoadingCharacters, setIsLoadingCharacters] = useState(false);
-  
-  // Refs
-  const fileInputRef = useRef<HTMLInputElement>(null);
-  const audioInputRef = useRef<HTMLInputElement>(null);
-  const videoRef = useRef<HTMLVideoElement>(null);
-  const canvasRef = useRef<HTMLCanvasElement>(null);
-  const mediaRecorderRef = useRef<MediaRecorder | null>(null);
-  const audioChunksRef = useRef<Blob[]>([]);
-  const saveTimeoutRef = useRef<NodeJS.Timeout | null>(null);
+  const {
+    storyText,
+    setStoryText,
+    uploadedImages,
+    setUploadedImages,
+    uploadedAudio,
+    setUploadedAudio,
+    isSaving,
+    saveToSession,
+  } = useStep2Session();
 
-  // Load from sessionStorage on mount
-  useEffect(() => {
-    const savedData = sessionStorage.getItem('step2Data');
-    if (savedData) {
-      try {
-        const data: SessionData = JSON.parse(savedData);
-        setStoryText(data.text || '');
-        // Note: We can't restore file objects from sessionStorage, only indicate they existed
-        if (data.images && data.images.length > 0) {
-          // Show indicator that images were previously uploaded
-          console.log('Previous images detected:', data.images.length);
-        }
-        if (data.audio) {
-          console.log('Previous audio detected');
-        }
-      } catch (error) {
-        console.error('Error loading saved data:', error);
-      }
-    }
-    
-    // Load existing characters
-    fetchExistingCharacters();
-  }, []);
-
-  // Fetch user's existing characters
-  const fetchExistingCharacters = async () => {
-    try {
-      setIsLoadingCharacters(true);
-
-      // Fetch characters for the current author
-      const charactersResponse = await fetch('/api/characters');
-      if (charactersResponse.ok) {
-        const data = await charactersResponse.json();
-        if (data.success && data.characters) {
-          setExistingCharacters(data.characters);
-        }
-      }
-    } catch (error) {
-      console.error('Error fetching existing characters:', error);
-    } finally {
-      setIsLoadingCharacters(false);
-    }
-  };
-
-  // Rotate tips
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setCurrentTipIndex((prev) => (prev + 1) % WRITING_TIPS.length);
-    }, 5000);
-    return () => clearInterval(interval);
-  }, []);
-
-  // Auto-save functionality with debouncing
-  const saveToSession = useCallback(() => {
-    setIsSaving(true);
-    const data: SessionData = {
-      text: storyText,
-      images: uploadedImages.map(img => img.preview),
-      audio: uploadedAudio?.preview || null,
-      lastSaved: Date.now()
-    };
-    sessionStorage.setItem('step2Data', JSON.stringify(data));
-    setTimeout(() => setIsSaving(false), 500);
-  }, [storyText, uploadedImages, uploadedAudio]);
-
-  // Debounced save on text change
-  useEffect(() => {
-    if (saveTimeoutRef.current) {
-      clearTimeout(saveTimeoutRef.current);
-    }
-    saveTimeoutRef.current = setTimeout(() => {
-      saveToSession();
-    }, 10000); // Save after 10 seconds of inactivity
-
-    return () => {
-      if (saveTimeoutRef.current) {
-        clearTimeout(saveTimeoutRef.current);
-      }
-    };
-  }, [storyText, saveToSession]);
-
-  const handleImageUpload = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const files = event.target.files;
-    if (files) {
-      const newImages = Array.from(files).slice(0, 3 - uploadedImages.length);
-      const imagePromises = newImages.map(file => {
-        return new Promise<{file: File, preview: string}>((resolve) => {
-          const reader = new FileReader();
-          reader.onload = (e) => {
-            resolve({
-              file,
-              preview: e.target?.result as string
-            });
-          };
-          reader.readAsDataURL(file);
-        });
-      });
-      
-      Promise.all(imagePromises).then(results => {
-        setUploadedImages(prev => [...prev, ...results].slice(0, 3));
-        saveToSession();
-      });
-    }
-  };
-
-  const handleAudioUpload = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const file = event.target.files?.[0];
-    if (file) {
-      const reader = new FileReader();
-      reader.onload = (e) => {
-        setUploadedAudio({
-          file,
-          preview: e.target?.result as string
-        });
-        saveToSession();
-      };
-      reader.readAsDataURL(file);
-    }
-  };
-
-  const startCamera = async () => {
-    try {
-      setIsCapturing(true);
-      const stream = await navigator.mediaDevices.getUserMedia({
-        video: { facingMode: 'environment' }
-      });
-
-      if (videoRef.current) {
-        videoRef.current.srcObject = stream;
-        videoRef.current.play();
-      }
-    } catch (error) {
-      console.error('Error accessing camera:', error);
-      alert(tStoryStepsStep2('alerts.cameraIssue'));
-      setIsCapturing(false);
-    }
-  };
-
-  const capturePhoto = () => {
-    if (videoRef.current && canvasRef.current && uploadedImages.length < 3) {
-      const video = videoRef.current;
-      const canvas = canvasRef.current;
-      const context = canvas.getContext('2d');
-
-      canvas.width = video.videoWidth;
-      canvas.height = video.videoHeight;
-
-      if (context) {
-        context.drawImage(video, 0, 0);
-        canvas.toBlob((blob) => {
-          if (blob) {
-            const file = new File([blob], `captured-photo-${Date.now()}.jpg`, { type: 'image/jpeg' });
-            const preview = canvas.toDataURL();
-            setUploadedImages(prev => [...prev, { file, preview }].slice(0, 3));
-            saveToSession();
-            if (uploadedImages.length < 2) {
-              // Continue capturing if less than 3 images
-            } else {
-              stopCamera();
-            }
-          }
-        }, 'image/jpeg', 0.8);
-      }
-    }
-  };
-
-  const stopCamera = () => {
-    if (videoRef.current && videoRef.current.srcObject) {
-      const stream = videoRef.current.srcObject as MediaStream;
-      stream.getTracks().forEach(track => track.stop());
-      videoRef.current.srcObject = null;
-    }
-    setIsCapturing(false);
-  };
-
-  const removeImage = (index: number) => {
-    setUploadedImages(prev => prev.filter((_, i) => i !== index));
-    saveToSession();
-  };
-
-  const clearAudio = () => {
-    setUploadedAudio(null);
-    if (audioInputRef.current) {
-      audioInputRef.current.value = '';
-    }
-    saveToSession();
-  };
-
-  // Character selection handlers
-  const toggleCharacterSelection = (characterId: string) => {
-    setSelectedCharacterIds(prev => 
-      prev.includes(characterId) 
-        ? prev.filter(id => id !== characterId)
-        : [...prev, characterId]
-    );
-  };
-
-  const toggleSelectAllCharacters = () => {
-    if (selectedCharacterIds.length === existingCharacters.length) {
-      setSelectedCharacterIds([]);
-    } else {
-      setSelectedCharacterIds(existingCharacters.map(char => char.characterId));
-    }
-  };
-
-  const startRecording = async () => {
-    try {
-      const stream = await navigator.mediaDevices.getUserMedia({ audio: true });
-      const mediaRecorder = new MediaRecorder(stream);
-      
-      mediaRecorderRef.current = mediaRecorder;
-      audioChunksRef.current = [];
-      
-      mediaRecorder.ondataavailable = (event) => {
-        if (event.data.size > 0) {
-          audioChunksRef.current.push(event.data);
-        }
-      };
-      
-      mediaRecorder.onstop = () => {
-        const audioBlob = new Blob(audioChunksRef.current, { type: 'audio/wav' });
-        const audioFile = new File([audioBlob], 'recorded-audio.wav', { type: 'audio/wav' });
-        const audioUrl = URL.createObjectURL(audioBlob);
-        
-        setUploadedAudio({
-          file: audioFile,
-          preview: audioUrl
-        });
-        saveToSession();
-        stream.getTracks().forEach(track => track.stop());
-      };
-      
-      mediaRecorder.start();
-      setIsRecording(true);
-    } catch (error) {
-      console.error('Error accessing microphone:', error);
-      alert(tStoryStepsStep2('alerts.microphoneIssue'));
-    }
-  };
-
-  // Helper: convert File to data URL (base64)
-  const fileToDataUrl = (file: File) => new Promise<string>((resolve, reject) => {
-    const reader = new FileReader();
-    reader.onload = () => resolve(reader.result as string);
-    reader.onerror = (e) => reject(e);
-    reader.readAsDataURL(file);
-  });
-
-  const stopRecording = () => {
-    if (mediaRecorderRef.current && isRecording) {
-      mediaRecorderRef.current.stop();
-      setIsRecording(false);
-    }
-  };
-
+  const { uploadMedia } = useMediaUpload();
   const handleNextStep = async () => {
     try {
       setIsCreatingStory(true);
@@ -367,47 +77,29 @@ export default function Step2Page() {
       const { story } = await response.json();
       localStorage.setItem('currentStoryId', story.storyId);
 
-      // Process with GenAI if any content provided
-      if (storyText.trim() || uploadedImages.length > 0 || uploadedAudio) {
-        console.log('Processing content with GenAI (signed uploads)...');
+        // Process with GenAI if any content provided
+        if (storyText.trim() || uploadedImages.length > 0 || uploadedAudio) {
+          console.log('Processing content with GenAI (signed uploads)...');
 
-  let imageObjectPath: string | undefined;
-  let audioObjectPath: string | undefined;
-  let imageDataB64: string | undefined;
-  let audioDataB64: string | undefined;
+          let imageObjectPath: string | undefined;
+          let audioObjectPath: string | undefined;
+          let imageDataB64: string | undefined;
+          let audioDataB64: string | undefined;
 
-        // Minimal integration: upload first image if available
-        if (uploadedImages.length > 0) {
-          const img = uploadedImages[0].file;
-          const contentType = img.type || 'image/jpeg';
-          imageDataB64 = await fileToDataUrl(img);
-          const resp = await fetch('/api/media/signed-upload', {
-            method: 'POST',
-            headers: { 'Content-Type': 'application/json' },
-            body: JSON.stringify({ storyId: story.storyId, kind: 'image', contentType, filename: img.name, dataUrl: imageDataB64 })
-          });
-          const uploaded = await resp.json();
-          if (!resp.ok || !uploaded?.objectPath) throw new Error(`Image upload failed: ${resp.status}`);
-          imageObjectPath = uploaded.objectPath;
-        }
+          if (uploadedImages.length > 0) {
+            const { objectPath, dataUrl } = await uploadMedia(story.storyId, 'image', uploadedImages[0].file);
+            imageObjectPath = objectPath;
+            imageDataB64 = dataUrl;
+          }
 
-        // Upload audio if present
-        if (uploadedAudio) {
-          const aud = uploadedAudio.file;
-          const contentType = aud.type || 'audio/wav';
-          audioDataB64 = await fileToDataUrl(aud);
-          const resp = await fetch('/api/media/signed-upload', {
-            method: 'POST',
-            headers: { 'Content-Type': 'application/json' },
-            body: JSON.stringify({ storyId: story.storyId, kind: 'audio', contentType, filename: aud.name, dataUrl: audioDataB64 })
-          });
-          const uploaded = await resp.json();
-          if (!resp.ok || !uploaded?.objectPath) throw new Error(`Audio upload failed: ${resp.status}`);
-          audioObjectPath = uploaded.objectPath;
-        }
+          if (uploadedAudio) {
+            const { objectPath, dataUrl } = await uploadMedia(story.storyId, 'audio', uploadedAudio.file);
+            audioObjectPath = objectPath;
+            audioDataB64 = dataUrl;
+          }
 
-        // Send to GenAI for processing with object paths
-    const genaiResponse = await fetch('/api/stories/genai-structure', {
+          // Send to GenAI for processing with object paths
+          const genaiResponse = await fetch('/api/stories/genai-structure', {
           method: 'POST',
           headers: {
             'Content-Type': 'application/json',
@@ -462,7 +154,6 @@ export default function Step2Page() {
     }
   };
 
-  const currentTip = WRITING_TIPS[currentTipIndex];
 
   return (
     <>
@@ -563,83 +254,7 @@ export default function Step2Page() {
                   </button>
                 </div>
 
-                {/* Character Selection - Only show if user has characters */}
-                {existingCharacters.length > 0 && (
-                  <div className="bg-gray-50 border border-gray-200 rounded-lg">
-                    <div
-                      className="p-4 cursor-pointer flex items-center justify-between"
-                      onClick={() => setShowCharacterSelection(!showCharacterSelection)}
-                    >
-                      <div className="flex items-center gap-3">
-                        <span className="text-xl">👥</span>
-                        <div>
-                          <h3 className="font-semibold text-gray-800">{tStoryStepsStep2('characterSelection.includeExistingTitle')}</h3>
-                          <p className="text-sm text-gray-600">
-                            {selectedCharacterIds.length > 0
-                              ? tStoryStepsStep2('characterSelection.selectedCount', { count: selectedCharacterIds.length })
-                              : tStoryStepsStep2('characterSelection.noneSelected')}
-                          </p>
-                        </div>
-                      </div>
-                      <div className={`transform transition-transform ${showCharacterSelection ? 'rotate-180' : ''}`}>
-                        <svg className="w-5 h-5 text-gray-500" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                          <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 9l-7 7-7-7" />
-                        </svg>
-                      </div>
-                    </div>
-
-                    {showCharacterSelection && (
-                      <div className="border-t border-gray-200 p-4 max-h-60 overflow-y-auto">
-                        {isLoadingCharacters ? (
-                          <div className="flex items-center justify-center py-8">
-                            <span className="loading loading-spinner loading-md"></span>
-                            <span className="ml-2 text-gray-600">{tStoryStepsStep2('characterSelection.loading')}</span>
-                          </div>
-                        ) : (
-                          <>
-                            {existingCharacters.length > 1 && (
-                              <div className="mb-3 pb-3 border-b border-gray-200">
-                                <label className="flex items-center gap-2 cursor-pointer">
-                                  <input
-                                    type="checkbox"
-                                    className="checkbox checkbox-sm"
-                                    checked={selectedCharacterIds.length === existingCharacters.length}
-                                    onChange={toggleSelectAllCharacters}
-                                  />
-                                  <span className="text-sm font-medium text-gray-700">
-                                    {tStoryStepsStep2('characterSelection.selectAll', { count: existingCharacters.length })}
-                                  </span>
-                                </label>
-                              </div>
-                            )}
-
-                            <div className="space-y-2">
-                              {existingCharacters.map((character) => (
-                                <label key={character.characterId} className="flex items-center gap-3 p-2 rounded-lg hover:bg-gray-100 cursor-pointer">
-                                  <input
-                                    type="checkbox"
-                                    className="checkbox checkbox-sm"
-                                    checked={selectedCharacterIds.includes(character.characterId)}
-                                    onChange={() => toggleCharacterSelection(character.characterId)}
-                                  />
-                                  <div className="flex-1">
-                                    <div className="font-medium text-gray-800">{character.name}</div>
-                                    {(character.type || character.role) && (
-                                      <div className="text-xs text-gray-500">
-                                        {[character.type, character.role].filter(Boolean).join(' • ')}
-                                      </div>
-                                    )}
-                                  </div>
-                                </label>
-                              ))}
-                            </div>
-                          </>
-                        )}
-                      </div>
-                    )}
-                  </div>
-                )}
-
+                <CharacterSelection onChange={setSelectedCharacterIds} />
                 {/* Reassurance Message */}
                 <div className="bg-blue-50 border border-blue-200 rounded-lg p-4">
                   <div className="flex items-start space-x-3">
@@ -762,17 +377,9 @@ export default function Step2Page() {
                   </label>
                 </div>
 
-                {/* Writing Tips - Moved to bottom */}
+                {/* Writing Tips */}
                 <div className="px-6 pb-4">
-                  <div className="p-4 bg-base-200 rounded-lg">
-                    <h4 className="font-semibold mb-2 flex items-center gap-2">
-                      <span className="text-xl">{currentTip.icon}</span>
-                      <span className="text-sm">{tStoryStepsStep2('writingTipsTitle')}</span>
-                    </h4>
-                    <p className="text-sm leading-relaxed animate-fade-in">
-                      {tStoryStepsStep2(`writingTips.${currentTipIndex}.text`)}
-                    </p>
-                  </div>
+                  <WritingTips />
                 </div>
               </div>
 
@@ -791,266 +398,16 @@ export default function Step2Page() {
           </div>
         )}
 
-        {/* Image Modal */}
-        {activeModal === 'images' && (
-          <div className="modal modal-open">
-            <div className="modal-box max-w-5xl w-11/12 h-[90vh] flex flex-col">
-              <div className="modal-header flex justify-between items-center mb-4">
-                <h3 className="font-bold text-2xl">📸 {tStoryStepsStep2('tabImage')}</h3>
-                <button
-                  className="btn btn-sm btn-circle btn-ghost"
-                  onClick={() => {
-                    setActiveModal(null);
-                    stopCamera();
-                  }}
-                >
-                  ✕
-                </button>
-              </div>
-              
-              <div className="flex-1 overflow-y-auto">
-                {/* Image Gallery */}
-                {uploadedImages.length > 0 && (
-                  <div className="mb-6">
-                    <h4 className="font-semibold mb-3">
-                      {tStoryStepsStep2('imageGalleryTitle', { count: uploadedImages.length })}
-                    </h4>
-                    <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
-                      {uploadedImages.map((img, index) => (
-                        <div key={index} className="relative">
-                          <div className="aspect-video rounded-lg overflow-hidden bg-base-200">
-                            <Image
-                              src={img.preview}
-                              alt={`Story image ${index + 1}`}
-                              fill
-                              className="object-cover"
-                            />
-                          </div>
-                          <button
-                            className="btn btn-sm btn-circle btn-error absolute top-2 right-2"
-                            onClick={() => removeImage(index)}
-                          >
-                            ✕
-                          </button>
-                        </div>
-                      ))}
-                    </div>
-                  </div>
-                )}
+        <MediaCapture
+          activeModal={activeModal === 'images' || activeModal === 'audio' ? activeModal : null}
+          setActiveModal={setActiveModal}
+          uploadedImages={uploadedImages}
+          setUploadedImages={setUploadedImages}
+          uploadedAudio={uploadedAudio}
+          setUploadedAudio={setUploadedAudio}
+          saveToSession={saveToSession}
+        />
 
-                {/* Add Images Section */}
-                {uploadedImages.length < 3 && !isCapturing && (
-                  <div className="text-center space-y-4">
-                    <div className="flex flex-col sm:flex-row gap-4 justify-center">
-                      <button
-                        className="btn btn-primary btn-lg"
-                        onClick={startCamera}
-                      >
-                        📷 {tStoryStepsStep2('takePhoto')}
-                      </button>
-                      <button
-                        className="btn btn-outline btn-lg"
-                        onClick={() => fileInputRef.current?.click()}
-                      >
-                        🖼️ {tStoryStepsStep2('uploadImage')}
-                      </button>
-                    </div>
-                    <p className="text-gray-600">{tStoryStepsStep2('imageHelp')}</p>
-                  </div>
-                )}
-
-                {/* Camera View */}
-                {isCapturing && (
-                  <div className="text-center space-y-4">
-                    <video
-                      ref={videoRef}
-                      className="w-full max-w-md mx-auto rounded-lg border"
-                      autoPlay
-                      playsInline
-                      muted
-                    />
-                    <div className="flex gap-4 justify-center">
-                      <button
-                        className="btn btn-primary btn-lg"
-                        onClick={capturePhoto}
-                        disabled={uploadedImages.length >= 3}
-                      >
-                        {tStoryStepsStep2('buttons.capture')}
-                      </button>
-                      <button
-                        className="btn btn-outline btn-lg"
-                        onClick={stopCamera}
-                      >
-                        {tStoryStepsStep2('buttons.cancel')}
-                      </button>
-                    </div>
-                  </div>
-                )}
-
-                <input
-                  ref={fileInputRef}
-                  type="file"
-                  accept="image/*"
-                  onChange={handleImageUpload}
-                  multiple
-                  className="hidden"
-                />
-                <canvas ref={canvasRef} className="hidden" />
-
-                {/* Image Tips */}
-                <div className="mt-6 p-4 bg-base-200 rounded-lg">
-                  <h4 className="font-semibold mb-3">{tStoryStepsStep2('photoTips.title')}</h4>
-                  <ul className="text-sm space-y-1 list-disc list-inside">
-                    {(tStoryStepsStep2.raw('photoTips.tips') as string[]).map((tip: string, index: number) => (
-                      <li key={index}>{tip}</li>
-                    ))}
-                  </ul>
-                </div>
-              </div>
-
-              <div className="modal-action">
-                <button
-                  className="btn btn-primary"
-                  onClick={() => {
-                    setActiveModal(null);
-                    stopCamera();
-                    saveToSession();
-                  }}
-                >
-                  {tStoryStepsStep2('actions.done')}
-                </button>
-              </div>
-            </div>
-          </div>
-        )}
-
-        {/* Audio Modal */}
-        {activeModal === 'audio' && (
-          <div className="modal modal-open">
-            <div className="modal-box max-w-5xl w-11/12 h-[90vh] flex flex-col">
-              <div className="modal-header flex justify-between items-center mb-4">
-                <h3 className="font-bold text-2xl">🎤 {tStoryStepsStep2('tabRecord')}</h3>
-                <button
-                  className="btn btn-sm btn-circle btn-ghost"
-                  onClick={() => setActiveModal(null)}
-                >
-                  ✕
-                </button>
-              </div>
-              
-              <div className="flex-1 overflow-y-auto">
-                {!uploadedAudio && !isRecording && (
-                  <div className="text-center space-y-4">
-                    <div className="flex flex-col sm:flex-row gap-4 justify-center">
-                      <button
-                        className="btn btn-primary btn-lg"
-                        onClick={startRecording}
-                      >
-                        🎤 {tStoryStepsStep2('recordVoice')}
-                      </button>
-                      <button
-                        className="btn btn-outline btn-lg"
-                        onClick={() => audioInputRef.current?.click()}
-                      >
-                        📁 {tStoryStepsStep2('uploadAudio')}
-                      </button>
-                    </div>
-                    <p className="text-gray-600">{tStoryStepsStep2('audioHelp')}</p>
-                  </div>
-                )}
-
-                {/* Recording View */}
-                {isRecording && (
-                  <div className="text-center space-y-4">
-                    <div className="flex flex-col items-center space-y-4">
-                      <div className="relative">
-                        <div className="w-32 h-32 bg-red-500 rounded-full flex items-center justify-center animate-pulse">
-                          <div className="text-white text-4xl">🎤</div>
-                        </div>
-                        <div className="absolute inset-0 rounded-full border-4 border-red-500 animate-ping"></div>
-                      </div>
-                      <p className="text-lg font-semibold text-red-600">{tStoryStepsStep2('actions.recording')}</p>
-                      <p className="text-gray-600">{tStoryStepsStep2('recordingHelp')}</p>
-                    </div>
-                    <div className="flex gap-4 justify-center">
-                      <button
-                        className="btn btn-error btn-lg"
-                        onClick={stopRecording}
-                      >
-                        ⏹️ {tStoryStepsStep2('actions.stopRecording')}
-                      </button>
-                    </div>
-                  </div>
-                )}
-
-                {/* Audio Preview */}
-                {uploadedAudio && (
-                  <div className="text-center space-y-4">
-                    <div className="card bg-base-200">
-                      <div className="card-body">
-                        <div className="flex items-center justify-center mb-4">
-                          <div className="text-6xl">🎵</div>
-                        </div>
-                        <audio
-                          src={uploadedAudio.preview}
-                          controls
-                          className="w-full max-w-md mx-auto"
-                        >
-                          {tStoryStepsStep2('audioSupport.notSupported')}
-                        </audio>
-                        <div className="card-actions justify-center">
-                          <button
-                            className="btn btn-outline btn-sm"
-                            onClick={clearAudio}
-                          >
-                            🗑️ {tStoryStepsStep2('actions.remove')}
-                          </button>
-                          <button
-                            className="btn btn-primary btn-sm"
-                            onClick={startRecording}
-                          >
-                            🔄 {tStoryStepsStep2('actions.recordAgain')}
-                          </button>
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                )}
-
-                <input
-                  ref={audioInputRef}
-                  type="file"
-                  accept="audio/mp3,audio/mpeg,audio/wav,audio/m4a"
-                  onChange={handleAudioUpload}
-                  className="hidden"
-                />
-
-                {/* Audio Tips */}
-                <div className="mt-6 p-4 bg-base-200 rounded-lg">
-                  <h4 className="font-semibold mb-3">🎤 {tStoryStepsStep2('recordingTips.title')}</h4>
-                  <ul className="text-sm space-y-1 list-disc list-inside">
-                    {(tStoryStepsStep2.raw('recordingTips.tips') as string[]).map((tip: string, index: number) => (
-                      <li key={index}>{tip}</li>
-                    ))}
-                  </ul>
-                </div>
-              </div>
-
-              <div className="modal-action">
-                <button
-                  className="btn btn-primary"
-                  onClick={() => {
-                    setActiveModal(null);
-                    saveToSession();
-                  }}
-                >
-                  {tStoryStepsStep2('actions.done')}
-                </button>
-              </div>
-            </div>
-          </div>
-        )}
-        
         {/* Loading Modal */}
         {showLoadingModal && (
           <div className="modal modal-open">
